fixed warning from heap test
git-svn-id: https://svn.code.sf.net/p/peano/code/trunk@3189 aa977e5b-e928-4c84-aa66-c90e40304b12

+++ src/peano/heap/tests/AggregationBoundaryDataExchangerTest.cpp
@@ -128,8 +128,8 @@ void peano::heap::tests::AggregationBoundaryDataExchangerTest::testComposeDecomp
     validateEqualsWithParams3(
       sentNumberOfMessgaes,i,
 	  exchanger._numberOfSentMessages,
-	  (int)(std::numeric_limits<double>::max()),
-	  (int)(exchanger._aggregatedSendData[0])
+	  std::numeric_limits<double>::max(),
+	  exchanger._aggregatedSendData[0]
 	);
 
     receivedTask.freeMemory();
